Add node tests
