+++ package.json
@@ -4,7 +4,7 @@
   "description": "Web game, based on The Resistance - social deduction",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node --test"
   },
   "repository": {
     "type": "git",
